% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourPHFfit.R
\name{FourPHF}
\alias{FourPHF}
\alias{FourPHF_fixa}
\alias{FourPHF_fixy0}
\alias{FourPHF_fixa_fixy0}
\alias{RateofGerm}
\title{Four paramter hill function}
\usage{
FourPHF(x, a, bta, c, y0)

FourPHF_fixa(x, a = 100, bta, c, y0)

FourPHF_fixy0(x, a, bta, c)

FourPHF_fixa_fixy0(x, a = 100, bta, c)

RateofGerm(x, a, b, c)
}
\arguments{
\item{x}{The explanatory/independent variable value.}

\item{a}{Parameter \mjseqn{a}.}

\item{bta}{Parameter \mjseqn{\beta} where \mjseqn{b = e^{\beta}}.}

\item{c}{Parameter \mjseqn{c}.}

\item{y0}{Parameter \mjseqn{y_{0}}.}

\item{b}{Parameter \mjseqn{b}.}
}
\value{
The calculated response/dependent value value.
}
\description{
To be used by
\code{\link[germinationmetrics]{FourPHFfit}}.
}
