% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcdata.R
\docType{data}
\name{gcdata}
\alias{gcdata}
\title{Germination Count Data}
\format{A data frame with 17 columns: \describe{ \item{Genotype}{The
  genotype.} \item{Rep}{Replication.} \item{Day1 to Day14}{Number of seeds
  that germinated (germination counts) on a particular day.} \item{Total
  Seeds}{Total number of seeds tested.} }}
\usage{
gcdata
}
\description{
An example germination count dataset. It includes germination count data over
14 days from five genotypes (G1 to G5) in three replications
}
\examples{

}
\keyword{datasets}
