% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germination.indices.R
\name{germination.indices}
\alias{germination.indices}
\title{Germination Indices}
\usage{
germination.indices(data, total.seeds.col, counts.intervals.cols,
  intervals, partial = TRUE, FirstGermTime = TRUE,
  LastGermTime = TRUE, PeakGermTime = TRUE, TimeSpreadGerm = TRUE,
  t50 = TRUE, MeanGermTime = TRUE, VarGermTime = TRUE,
  SEGermTime = TRUE, CVSEGermTime = TRUE, MeanGermRate = TRUE,
  CVG = TRUE, VarGermRate = TRUE, SEGermRate = TRUE,
  GermRateRecip = TRUE, GermSpeed = TRUE,
  GermSpeedAccumulated = TRUE, GermSpeedCorrected = TRUE,
  MeanGermPercent = TRUE, MeanGermNumber = TRUE, TimsonsIndex = TRUE,
  GermRateGeorge = TRUE, PeakValue = TRUE, GermValue = TRUE,
  CUGerm = TRUE, GermSynchrony = TRUE, GermUncertainty = TRUE)
}
\arguments{
\item{data}{A data frame with the germination count data. It should possess
columns with \itemize{ \item Partial or cumulative germination counts per
time interval (to be indicated by the argument \code{counts.intervals.cols}
and \item Total number of seeds tested (to be indicated by the argument
\code{total.seeds.col}). }}

\item{total.seeds.col}{The name of the column with the total number of seeds
tested.}

\item{counts.intervals.cols}{The names of columns with the germination counts
(partial or cumulative, as indicated by the argument \code{partial}) per
time interval (indicated by the argument \code{intervals}).}

\item{intervals}{The time intervals at which germination counts (in the
columns specified in argument \code{counts.intervals.cols})  were recorded.}

\item{partial}{logical. If \code{TRUE}, germination counts in
\code{counts.intervals.cols} is considered as partial and if \code{FALSE},
it is considered as cumulative. Default is \code{TRUE}.}
}
\value{
A data frame with the original data along with the various
  single-value germination indices specified in the function arguments.
}
\description{
Compute the following germination indices from a data frame of germination
counts recorded at specific time intervals for multiple samples in batch.
\itemize{ \item Mean germination time (\ifelse{html}{\out{<em><span
style="text-decoration:overline">T</span></em>}}{\eqn{\overline{T}}}) or Mean
length of incubation time
\insertCite{edmond_effects_1958,czabator_germination_1962,ellis_improved_1980,labouriau_germinacao_1983,ranal_how_2006}{germinationmetrics}.
\item Variance of germination time (\ifelse{html}{\out{<em>s<sup>2</sup><sub
style="line-height: 1.8; margin-left: -1ex;">T</sub></em>}}{\eqn{s_{T}^{2}}})
\insertCite{labouriau_germinacao_1983,ranal_how_2006}{germinationmetrics}.
\item Standard error of germination time (\ifelse{html}{\out{<em>s<sub><span
style="text-decoration:overline">T</span></sub></em>}}{\eqn{s_{\overline{T}}}})
\insertCite{labouriau_germinacao_1983,ranal_how_2006}{germinationmetrics}.
\item Coefficient of variation of the germination time
(\ifelse{html}{\out{<i>CV<sub>T</sub></i>}}{\eqn{CV_{T}}})
\insertCite{ranal_how_2006}{germinationmetrics}. }
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
This function is a wrapper around the different functions for
  computation of single-value germination indices in
  \href{https://cran.r-project.org/package=germinationmetrics}{germinationmetrics}
   (\code{\link[germinationmetrics]{MeanGermTime}},
  \code{\link[germinationmetrics]{t50}})
}
